corrector_project/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
arrow==1.2.3
    # via -r requirements.in
asgiref==3.6.0
    # via django
click==8.1.3
    # via uvicorn
crispy-bootstrap5==0.7
    # via -r requirements.in
django==4.1.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-crispy-forms==1.14.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.9.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
psycopg2-binary==2.9.5
    # via -r requirements.in
python-dateutil==2.8.2
    # via arrow
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via django
uvicorn==0.21.1
    # via -r requirements.in
whitenoise==6.4.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
